Ignore the @line.max_length when determining whether or not to print a character.
The max_length is being used to stop input from going past the end of the current terminal row's width.

This introduces an issue if you try to go back to a previous line for editing, but that should be addressed separately. For now, assume I want rawline to accept any amount of user input.

Also, Line shouldn't enforce max_length when incrementing right.