HistoryBuffer#get now returns a copy of the element.
This is to preserve the text in the editor's history if you start modifying a line that you found while navigating history. 

You don't want the old historical version of a line to change, but a new line to be added if the user modifies the command (and they run the line by hitting enter or equivalent).