Adding Editor#prompt/prompt= accessor methods for setting the prompt text.
Also added a @allow_prompt_updates flag to determine when prompt updates are allowed. This is a poor man's solution for ensuring a multi-threaded program can update the prompt. 

Basically, don't allow updating if we're about to present a new line of input to the user. Not pretty or perfect, but works for now.