RawLine::EventLoop#add_event now possesses a little bit of intelligence.
If the last event queued is the same as the incoming one being queued then there is the incoming event can be discarded.

Example: If you copy and paste text into the editor it will process a change event for every character added which in turn propagates N render events for the N character(s). We don't need to render N times in a row, so just render once. 

AFAICT this applies to all events which are identical for whose use-cases I am currently considering.