Exploring using an environment stack for the editor.
The goal of this would be to allow for a better way to 
implement modes by allowing modes to completely own 
their environment. 

When the mode is activated (e.g. 
by a key binding) it would push its environment on
the editor's environment stack. This would allow it
to provide its own set of overriding environment 
characteristics. When mode was done its environment
would pop itself off the stack restoring the original 
environment.

It is still up for consideration whether if things
like key binding lookups should fall walk the environment
stack until it finds a match.