KeycodeParser has learned to parse byte sequences from the front going forward.
It's previous implementation searched from the back going backwards, but this takes far too long – seemingly locking up the console – when a user copies and pastes a long command.

The forward parsing is able to peek ahead to see if the current bytes we're looking at are the start of any known byte sequences. If they are not, it stops looking any further.