Etching in an Environment to support falling back to its parent env.
This pushes the ownership of the keyboard input processor to the Environment rather than having the Editor access env.keyboard_input_processors directly.

Related, any implementation of #read_bytes should now return nil or an empty array to indicate that it has no bytes to ask a parent env to process. If there are leftover bytes it  should return those and those will be bubbled.

On another note, this removes the need for tab completion to handle leftover bytes on its own and allows for a more generic (and configurable) approach for implementing modes.